requirements.txt
beautifulsoup4==4.12.3
dateparser==1.2.0
GitPython==3.1.43
pandas==2.2.2
python-slugify==8.0.4
pytz==2024.1
Requests==2.32.3
requests_cache==1.2.1
thefuzz==0.22.1
